+++ src/admin/documentTypes/document-type.controller.ts
@@ -1,30 +1,30 @@
 import { PfxHttpResponseDto } from 'profaxnojs/axios';
 import { SearchPaginationDto } from 'profaxnojs/util';
 
-import { Controller, Get, Body, Patch, Param, Logger, HttpCode, HttpStatus, Query, ParseUUIDPipe, NotFoundException } from '@nestjs/common';
+import { Controller, Get, Body, Patch, Param, Logger, HttpCode, HttpStatus, Query, ParseUUIDPipe, NotFoundException, Delete } from '@nestjs/common';
 
-import { ProductUnitDto, ProductUnitSearchInputDto } from './dto';
-import { ProductUnitService } from './product-unit.service';
+import { DocumentTypeDto, DocumentTypeSearchInputDto } from './dto';
+import { DocumentTypeService } from './document-type.service';
 
-import { AlreadyExistException } from '../../../common/exceptions/common.exception';
+import { AlreadyExistException, IsBeingUsedException } from '../../common/exceptions/common.exception';
 
-@Controller('setting-product-units')
-export class ProductUnitController {
+@Controller('document-types')
+export class DocumentTypeController {
 
-  private readonly logger = new Logger(ProductUnitController.name);
+  private readonly logger = new Logger(DocumentTypeController.name);
 
   constructor(
-    private readonly documentTypeService: ProductUnitService
+    private readonly documentTypeService: DocumentTypeService
   ) {}
 
   @Patch('/update')
   @HttpCode(HttpStatus.OK)
-  update(@Body() dto: ProductUnitDto): Promise<PfxHttpResponseDto> {
+  update(@Body() dto: DocumentTypeDto): Promise<PfxHttpResponseDto> {
     this.logger.log(`>>> update: dto=${JSON.stringify(dto)}`);
     const start = performance.now();
 
     return this.documentTypeService.update(dto)
-    .then( (dto: ProductUnitDto) => {
+    .then( (dto: DocumentTypeDto) => {
       const response = new PfxHttpResponseDto(HttpStatus.OK, 'executed', 1, [dto]);
       const end = performance.now();
       this.logger.log(`<<< update: executed, runtime=${(end - start) / 1000} seconds, response=${JSON.stringify(response)}`);
@@ -46,14 +46,14 @@ export class ProductUnitController {
   searchByValues(
     @Param('companyId', ParseUUIDPipe) companyId: string,
     @Query() paginationDto: SearchPaginationDto,
-    @Body() inputDto: ProductUnitSearchInputDto
+    @Body() inputDto: DocumentTypeSearchInputDto
   ): Promise<PfxHttpResponseDto> {
 
     this.logger.log(`>>> searchByValues: companyId=${companyId}, paginationDto=${JSON.stringify(paginationDto)}, inputDto=${JSON.stringify(inputDto)}`);
     const start = performance.now();
     
     return this.documentTypeService.searchByValues(companyId, paginationDto, inputDto)
-    .then( (dtoList: ProductUnitDto[]) => {
+    .then( (dtoList: DocumentTypeDto[]) => {
       const response = new PfxHttpResponseDto(HttpStatus.OK, "executed", dtoList.length, dtoList);
       const end = performance.now();
       this.logger.log(`<<< searchByValues: executed, runtime=${(end - start) / 1000} seconds, response=${JSON.stringify(response)}`);
@@ -62,6 +62,30 @@ export class ProductUnitController {
     .catch( (error: Error) => {
       if(error instanceof NotFoundException)
         return new PfxHttpResponseDto(HttpStatus.NOT_FOUND, error.message, 0, []);
+
+      this.logger.error(error.stack);
+      return new PfxHttpResponseDto(HttpStatus.INTERNAL_SERVER_ERROR, error.message);
+    })
+  }
+
+  @Delete('/:id')
+  remove(@Param('id', ParseUUIDPipe) id: string): Promise<PfxHttpResponseDto> {
+    this.logger.log(`>>> remove: id=${id}`);
+    const start = performance.now();
+
+    return this.documentTypeService.remove(id)
+    .then( (msg: string) => {
+      const response = new PfxHttpResponseDto(HttpStatus.OK, msg);
+      const end = performance.now();
+      this.logger.log(`<<< remove: executed, runtime=${(end - start) / 1000} seconds, response=${JSON.stringify(response)}`);
+      return response;
+    })
+    .catch( (error: Error) => {
+      if(error instanceof NotFoundException)
+        return new PfxHttpResponseDto(HttpStatus.NOT_FOUND, error.message, 0, []);
+
+      if(error instanceof IsBeingUsedException)
+        return new PfxHttpResponseDto(HttpStatus.BAD_REQUEST, error.message, 0, []);
 
       this.logger.error(error.stack);
       return new PfxHttpResponseDto(HttpStatus.INTERNAL_SERVER_ERROR, error.message);
