--- src/admin/settings/productUnits/dto/index.ts
@@ -1,2 +1,0 @@
-export { ProductUnitDto } from './product-unit.dto';
-export { ProductUnitSearchInputDto } from './product-unit-search.dto';
