--- src/admin/settings/productUnits/dto/product-unit-search.dto.ts
@@ -1,13 +1,0 @@
-import { IsArray, IsOptional, IsString } from "class-validator";
-
-export class ProductUnitSearchInputDto {
-  
-  @IsOptional()
-  @IsString()
-  name?: string;
-  
-  constructor(name?: string) {
-    this.name = name;
-  }
-
-}
