--- src/admin/settings/productUnits/entities/product-unit.entity.ts
@@ -1,23 +1,0 @@
-import { Company } from "src/admin/companies/entities/company.entity";
-import { Column, Entity, ManyToOne, OneToMany, PrimaryGeneratedColumn } from "typeorm";
-
-@Entity("adm_setting_product_unit")
-export class ProductUnit {
-  
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column('varchar', { length: 50 })
-  name: string;
-
-  @Column('boolean', { default: true })
-  active: boolean
-
-  @ManyToOne(
-    () => Company,
-    (company) => company.documentType,
-    { eager: true }
-  )
-  company: Company;
-
-}
